Help boost find the Python 3 include files
Boost does not correctly locate the python include files if the
architecture scheme from PE-3149 is used. It assumes they are
in include/python3.5 but they might be in include/python3.5m.
We therefore find the include directories ourselves using Python
and then update the project-config.jam file to include that
location.

Just in case, also check that boost_python library exists before trying
to modify it. In Boost 1.60 it seems that libboost_python does not
directly link against libpython so this may not be needed.